requirements.txt
addict==2.4.0
elasticdeform==0.4.9
geopandas==0.10.2
matplotlib==3.5.1
numpy==1.22.3
PyYAML==6.0
rasterio==1.2.10
scipy==1.8.0
segmentation_models_pytorch==0.2.1
Shapely==1.8.1.post1
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.0
tensorboard==2.4.1
segmentation-models-pytorch==0.2.1
scikit-image==0.19.2